requirements.txt
requests
pandas
beautifulsoup4
cairosvg
Pillow
tqdm
